move it to intersect.ts
