+++ tests/new-ui/raycast.test.ts
@@ -1,6 +1,6 @@
 import { Vec3 } from "../../cocos/core";
 import { Plane, Ray } from "../../cocos/core/geometry";
-import { UIElement } from "../../cocos/new-ui/base/ui-element";
+import { raycastPlane } from "../../cocos/core/geometry/intersect";
 
 test('raycastPlane', () => {
     // default as a xoz plane
@@ -13,7 +13,7 @@ test('raycastPlane', () => {
     const result: Vec3 = new Vec3();
 
     // 1.ray and plane coincide(fail)
-    UIElement.raycastPlane(result, ray, plane);
+    raycastPlane(result, ray, plane);
     expect(result.x).toBe(0);
     expect(result.y).toBe(0);
     expect(result.z).toBe(0);
@@ -24,7 +24,7 @@ test('raycastPlane', () => {
     pointC.set(1, 0, 1);
     Plane.fromPoints(plane, pointA, pointB, pointC);
     // ray is parallel to plane
-    UIElement.raycastPlane(result, ray, plane);
+    raycastPlane(result, ray, plane);
     expect(result.x).toBe(0);
     expect(result.y).toBe(0);
     expect(result.z).toBe(0);
@@ -35,7 +35,7 @@ test('raycastPlane', () => {
     pointC.set(1, 0, 1);
     Plane.fromPoints(plane, pointA, pointB, pointC);
     // ray starts at (0,0,0), and its direction is to -z, so it cannot raycast the plane
-    UIElement.raycastPlane(result, ray, plane);
+    raycastPlane(result, ray, plane);
     expect(result.x).toBe(0);
     expect(result.y).toBe(0);
     expect(result.z).toBe(0);
@@ -46,35 +46,35 @@ test('raycastPlane', () => {
     pointC.set(1, 0, -1);
     Plane.fromPoints(plane, pointA, pointB, pointC);
     // it can raycast the plane
-    UIElement.raycastPlane(result, ray, plane);
+    raycastPlane(result, ray, plane);
     expect(result.x).toBe(0);
     expect(result.y).toBe(0);
     expect(result.z).toBe(-1);
 
     // 5.set ray direction as (0,1,-1)(success)
     ray.d.set(0, 1, -1);
-    UIElement.raycastPlane(result, ray, plane);
+    raycastPlane(result, ray, plane);
     expect(result.x).toBe(0);
     expect(result.y).toBe(1);
     expect(result.z).toBe(-1);
 
     // 6.set ray origin as (1,0,0)(success)
     ray.o.set(1,0,0);
-    UIElement.raycastPlane(result, ray, plane);
+    raycastPlane(result, ray, plane);
     expect(result.x).toBe(1);
     expect(result.y).toBe(1);
     expect(result.z).toBe(-1);
 
     // 7.set ray direction as (1,1,-1)(success)
     ray.d.set(1, 1, -1);
-    UIElement.raycastPlane(result, ray, plane);
+    raycastPlane(result, ray, plane);
     expect(result.x).toBe(2);
     expect(result.y).toBe(1);
     expect(result.z).toBe(-1);
 
     // 8.set ray direction as (1,1,1)(fail)
     ray.d.set(1, 1, 1);
-    UIElement.raycastPlane(result, ray, plane);
+    raycastPlane(result, ray, plane);
     expect(result.x).toBe(2);
     expect(result.y).toBe(1);
     expect(result.z).toBe(-1);
